% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateSequenceWindow}
\alias{calculateSequenceWindow}
\title{Append a window column to a given data frame}
\usage{
calculateSequenceWindow(dataframe, sitecol, window)
}
\arguments{
\item{dataframe}{data.frame with a uniprot column, and another numeric column with site}

\item{sitecol}{Name of the column containing the site information}

\item{window}{Half-width of window (\code{window} of \code{3} results in 7-aa windows)}
}
\description{
This function silently drops any sites that do not have enough neighbouring
 amino acids to make a complete window.
}
\examples{
df <- data.frame(uniprot='q14118',site=c(30,40,50))
   # 11-aa windows
   calculateSequenceWindow(df,'site',5)
   # 5-aa windows
   calculateSequenceWindow(df,'site',2)
}

