% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cddidToSuperfamily}
\alias{cddidToSuperfamily}
\title{Replace CDD identifiers with cluster identifiers}
\usage{
cddidToSuperfamily(cddids)
}
\arguments{
\item{cddids}{Vector of cddids to look up}
}
\value{
vector of identifiers
}
\description{
By default, methods on the CDD data sources use the
cddid as the identifier for a domain. We can use the
cluster identifier as an alternative if we wish to
group similar domains.
}
\details{
This method connects to the CDD to download the
family to superfamily links for the database, and
caches this file locally.
}

