% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertEntrezIds}
\alias{convertEntrezIds}
\title{Convert entrez ids to a data frame of uniprot, genename and entrez id}
\usage{
convertEntrezIds(organism = 9606, ids = c())
}
\description{
Convert entrez ids to a data frame of uniprot, genename and entrez id
}

