% Generated by roxygen2 (4.0.1): do not edit by hand
\name{downloadUniprotGOA}
\alias{downloadUniprotGOA}
\title{Download annotations from the UniprotGOA site}
\usage{
downloadUniprotGOA(organism = 9606)
}
\arguments{
\item{organism}{NCBI taxonomy id (e.g. 9606) to get terms for}
}
\value{
Data frame with gene id, go_id, Evidence and Ontology columns
}
\description{
Download annotations from the UniprotGOA site
}

