% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findCommonPathways.graphite}
\alias{findCommonPathways.graphite}
\title{Given sets of genes, find the common pathways that these genes belong to within Graphite}
\usage{
findCommonPathways.graphite(organism = 9606, max_pathway_size = 30, ...)
}
\arguments{
\item{organism}{NCBI taxonomy ID for the organism that is being studied}

\item{max_pathway_size}{Maximum size of the pathway that should be checked}

\item{...}{Lists of Entrez gene identifiers}
}
\value{
Data frame with columns pathway, db (indicating source DB) and the gene identifiers that are common to that pathway
}
\description{
Given sets of genes, find the common pathways that these genes belong to within Graphite
}
\examples{
findCommonPathways(organism=9606,max_pathway_size=1000,seta=c('1605'),setb=c('1605'))
}

