% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findCommonPathways}
\alias{findCommonPathways}
\title{Given sets of genes, find the common pathways that these genes belong to}
\usage{
findCommonPathways(organism = 9606, max_pathway_size = 80, ...)
}
\arguments{
\item{organism}{NCBI taxonomy ID for the organism that is being studied}

\item{max_pathway_size}{Maximum size of the pathway that should be checked}

\item{...}{Lists of Entrez gene identifiers}
}
\value{
Data frame with columns pathway, db (indicating source DB) and the gene identifiers that are common to that pathway
}
\description{
This looks within ReactomeDB, and Graphite (encompassing many other databases)
}
\examples{
findCommonPathways(organism=9606,max_pathway_size=1000,seta=c('1605'),setb=c('1605'))
}

