% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateGONetwork}
\alias{generateGONetwork}
\title{Plot a GO network for a given set of GO identifiers}
\usage{
generateGONetwork(goids, ancestors = 4)
}
\arguments{
\item{goids}{GO ids to use as the basis to build the network}

\item{ancestors}{Number of ancestors to add to the graph to establish more remote connections}
}
\value{
List containing two elements, plot: Containing the plot, and coords: Containing the co-ordinates of each protein
}
\description{
Plot a GO network for a given set of GO identifiers
}

