% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateLogoPlot}
\alias{generateLogoPlot}
\title{Generate a Berry plot for a set of sequences}
\usage{
generateLogoPlot(dataframe, windowcol, frequencies = c(), labels = T)
}
\arguments{
\item{dataframe}{Data frame containing at least one column (windowcol) that has the windows to plot}

\item{windowcol}{Column in the dataframe to look for window sequences}

\item{frequencies}{Amino acid frequencies - defaults to frequencies of amino acids in UniProt December 2013 release}

\item{labels}{Boolean flag as to whether to give a plot title with the number of rows / proteins used}
}
\value{
berry logo plot
}
\description{
Generate a Berry plot for a set of sequences
}

