% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateReactomeNetwork}
\alias{generateReactomeNetwork}
\title{Plot a Reactome network from a given pathway URL}
\usage{
generateReactomeNetwork(pathwayURL)
}
\arguments{
\item{pathwayURL}{URL to retrieve pathway from (http:// or file://)}
}
\value{
List containing two elements, plot: Containing the plot, and coords: Containing the co-ordinates of each protein
}
\description{
Plot a Reactome network from a given pathway URL
}

