% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateStringNetwork}
\alias{generateStringNetwork}
\title{Plot a String network from the given seed UniProt accession identifiers}
\usage{
generateStringNetwork(organism = 9606, accessions = c("q14118", "P24043"),
  get.neighbours = F, threshold = 700)
}
\arguments{
\item{organism}{NCBI taxonomy ID for which to retrieve the String network}

\item{accessions}{Vector of UniProt accessions to retrieve the String network for}

\item{get.neighbours}{Boolean flag (TRUE / FALSE) indicating if immediate neighbours of the
                        given accessions should also be added to the graph}

\item{threshold}{Minimum threshold for the combined score criteria for addition to the graph}
}
\value{
List containing two elements, plot: Containing the plot, and coords: Containing the co-ordinates of each protein
}
\description{
At a given evidence threshold, we can retrieve the String database relations, and add them
to a graph object, which is then rendered
}

