% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateVennDiagram}
\alias{generateVennDiagram}
\title{Generate Venn diagrams}
\usage{
generateVennDiagram(...)
}
\arguments{
\item{...}{Vectors to draw a venn diagram for}
}
\value{
ggplot object
}
\description{
Generate Venn diagrams
}
\examples{
generateVennDiagram(a=c(1,2,3),b=c(2,3,4),c=c(1,3))
}

