% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geom_text_background}
\alias{geom_text_background}
\title{Rendering of text labels with a roundrect background on them}
\usage{
geom_text_background(...)
}
\arguments{
\item{expand}{Ratio to expand the text label background by (defaults to 1.2)}

\item{bgcol}{Colour to use to draw the background roundrect}

\item{bgfill}{Fill colour to use for the background roundrect}

\item{bgalpha}{Alpha opacity for the background roundrect}
}
\description{
Works the same as geom_text, but you can set extra parameters on the background
}

