% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCddNames}
\alias{getCddNames}
\title{Retrieve the names for a vector of cdddids}
\usage{
getCddNames(cddids)
}
\arguments{
\item{cddids}{Vector of cddids to lookup names for}

\item{data.frame}{Domain names}
}
\description{
Names are retrieved from the CDD database for all domains
and a data frame with columns \code{dom}, \code{id}, \code{short}, \code{long} is returned
}

