% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCytosolic}
\alias{getCytosolic}
\title{Get a list of potentially cytosolic proteins from uniprot identifiers
Simple check that sees if there are only cytosolic related terms associated
with a given identifier}
\usage{
getCytosolic(organism, uniprots)
}
\arguments{
\item{organism}{NCBI taxonomy id (e.g. 9606) to get terms for}

\item{uniprots}{UniProt identifiers to check for cytosolic}
}
\value{
Vector of UniProt identifiers
}
\description{
Get a list of potentially cytosolic proteins from uniprot identifiers
Simple check that sees if there are only cytosolic related terms associated
with a given identifier
}

