% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGOEnrichment}
\alias{getGOEnrichment}
\title{Perform an enrichment analysis on a given set of uniprot ids}
\usage{
getGOEnrichment(organism = 9606, uniprots, query_ids = c(),
  universe = c(), ontology = "BP", direction = "over",
  supplemental.terms = F, conditional = TRUE)
}
\arguments{
\item{organism}{NCBI taxonomy identifier (e.g. 9606 for human)}

\item{uniprots}{Vector of uniprot accessions to perform enrichment upon}

\item{query_ids}{Optional Entrez gene ids to use if there are no uniprot identifiers}

\item{universe}{Vector of uniprot accessions to use as the universe. Defaults to all GO annotations for the organism}

\item{ontology}{Ontology to perform enrichment in ('BP','CC','MF')}

\item{direction}{Direction to check for enrichment ('over' or 'under' for over- and under-representation respectively)}

\item{supplemental.terms}{Flag to see if supplemental annotations should be retrieved
from UniprotGOA. Can also set to a dataframe containing the
terms.}

\item{conditional}{When set to TRUE, the enrichment will be conditional, taking into account relationships of terms}
}
\value{
Enrichment object
}
\description{
Perform an enrichment analysis on a given set of uniprot ids
}

