% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGOenrichmentGenes}
\alias{getGOenrichmentGenes}
\title{Get the genes that contributed to a particular enrichment of a set of terms}
\usage{
getGOenrichmentGenes(enrichment, wanted_terms = c(), organism = 9606)
}
\arguments{
\item{enrichment}{Enrichment obtained from \code{\link{getGOEnrichment}}}

\item{wanted_terms}{vector of go ids that you wish to link to enriched genes}

\item{organism}{Organism that you wish to search for terms in}
}
\value{
Data frame with uniprot id, gene id and gene name
}
\description{
Get the genes that contributed to a particular enrichment of a set of terms
}

