% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGONetwork}
\alias{getGONetwork}
\title{Create a GO network for a given set of GO identifiers}
\usage{
getGONetwork(goids, ancestors = 4)
}
\arguments{
\item{goids}{GO ids to use as the basis to build the network}

\item{ancestors}{Number of ancestors to add to the graph to establish more remote connections}
}
\value{
Network object with the GO network
}
\description{
Create a GO network for a given set of GO identifiers
}

