% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGOParents}
\alias{getGOParents}
\title{Get the parents of the given GO terms}
\usage{
getGOParents(node = "GO:0008150", ontology = "BP")
}
\arguments{
\item{node}{GO node to examine}

\item{ontology}{Ontology to look for parents in}
}
\description{
Get the parents of the given GO terms
}

