% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGOTerms}
\alias{getGOTerms}
\title{Get the GO terms associated with the given UniProt ids}
\usage{
getGOTerms(organism = 9606, uniprots, wanted = c(), ontology = "BP")
}
\arguments{
\item{organism}{Organism to look for GO terms for}

\item{uniprots}{Vector of uniprot ids to retrieve GO terms for}

\item{wanted}{Top-level GO terms to search under}

\item{ontology}{GO ontology to search ('BP','MF','CC')}
}
\description{
You can optionally supply a vector of GO terms to the method,
so that any children of the wanted terms will be considered as
the parent term. This is a cheap way of getting a slim GO set.
For example, if your wanted GO term was 'GO:0005829' (cytosol),
a gene annotated with the term 'GO:0044445' (cytosolic part)
will be mapped to 'GO:0005829' (cytosol).
}

