% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getReactomePathway}
\alias{getReactomePathway}
\title{Load a Reactome pathway network from an URL or local file}
\usage{
getReactomePathway(pathwayURL)
}
\arguments{
\item{pathwayURL}{URL to retrieve pathway from (http:// or file://)}
}
\value{
Network object for the given pathway
}
\description{
Load a Reactome pathway network from an URL or local file
}
\examples{
# Load from a local file
 getReactomePathway("file://mypathway.biopax")
 # Load from ReactomeDB
 url <- 'http://www.reactome.org/ReactomeRESTfulAPI/RESTfulWS/biopaxExporter/Level3/535734'
 network <- getReactomePathway(url)
}

