% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getStringNetwork}
\alias{getStringNetwork}
\title{Get a String network from the given seed UniProt accession identifiers}
\usage{
getStringNetwork(organism = 9606, accessions = c(), get.neighbours = F,
  threshold = 700)
}
\arguments{
\item{organism}{NCBI taxonomy ID for which to retrieve the String network}

\item{accessions}{Vector of UniProt accessions to retrieve the String network for}

\item{get.neighbours}{Boolean flag (TRUE / FALSE) indicating if immediate neighbours of the
                        given accessions should also be added to the graph}

\item{threshold}{Minimum threshold for the combined score criteria for addition to the graph}
}
\value{
A network class object
}
\description{
At a given evidence threshold, we can retrieve the String database relations, and add them
to a graph object. This graph can then be drawn using the \code{\link{ggnet}} method
}
\seealso{
\code{\link[network]{network}} \code{\link[STRINGdb]{STRINGdb}} \code{\link{ggnet}}
}

