% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getUniprotIds}
\alias{getUniprotIds}
\title{Retrieve the set of UniProt accessions for a given taxonomy ID}
\usage{
getUniprotIds(taxonomy)
}
\arguments{
\item{taxonomy}{NCBI taxonomy identifier to retrieve data for}
}
\value{
Vector of string accessions
}
\description{
Only get the identifiers for entries where they are part of a complete proteome set (keyword 1185)
and are reviewed
}

