% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getUniprotSequences}
\alias{getUniprotSequences}
\title{Retrieve UniProt sequences for a set of UniProt accessions}
\usage{
getUniprotSequences(accessions, wait = 0)
}
\arguments{
\item{accessions}{Uniprot accessions to retrieve sequence data for}
}
\value{
Data frame with UniProt identifier and sequence
}
\description{
Retrieve UniProt sequences for a set of UniProt accessions
}

