% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GOnetworkPlot.annotate}
\alias{GOnetworkPlot.annotate}
\title{Annotate a GO based network plot with data from a table of GO term datasource mappings}
\usage{
GOnetworkPlot.annotate(plot, annotations = data.frame(GOID = c(), datasource =
  c()), label.terms = F, ...)
}
\arguments{
\item{plot}{Result from \code{\link{generateGONetwork}}}

\item{annotations}{Data table with two columns - \code{GOID},\code{datasource}. Can also use GOBPID, GOCCID and GOMFID as GO identifier column}

\item{label.terms}{Toggle if the GO terms should be overlaid on the graph}

\item{...}{Colours to use for each of the given datasources}
}
\description{
Given a network plot obtained from  a GO network ( \code{\link{generateGONetwork}})
overlay annotation information as a set of halos.
}

