% Generated by roxygen2 (4.0.2): do not edit by hand
\name{networkPlot.annotate}
\alias{networkPlot.annotate}
\title{Annotate a protein based network plot with data from a table of annotations}
\usage{
networkPlot.annotate(organism = 9606, plot, annotations = data.frame(uniprot
  = NA, value = NA, datasource = NA), uniprot.symbols = F, node.color = NA,
  ...)
}
\arguments{
\item{organism}{NCBI taxonomy id for the organism that we are working with}

\item{plot}{Result from either \code{\link{generateStringNetwork}} or \code{\link{generateReactomeNetwork}}}

\item{annotations}{Data table with three columns - \code{uniprot},\code{value},\code{datasource}}

\item{uniprot.symbols}{Toggle if the uniprot symbols should be overlaid on the graph}

\item{node.color}{Datasource string value to use for setting the colour of the node}

\item{...}{Colour ranges to use for the scales for each of the given datasources}
}
\description{
Given a network plot obtained from String \code{\link{generateStringNetwork}} or Reactome \code{\link{generateReactomeNetwork}}
overlay expression information as a set of halos.
}

