% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{Rgator-deprecated}
\alias{GO_children}
\alias{GO_parents}
\alias{Rgator-deprecated}
\alias{Rgator-deprecated-package}
\alias{getDomainSets}
\alias{overlayExpression.protein}
\alias{overlayGOValues}
\alias{tableGOTerms}
\title{Deprecated functions}
\usage{
overlayExpression.protein(organism, plot, expression, uniprot.symbols,
  node.color, ...)

getDomainSets(inputsites, sitecol, domaindata, max_dom_proportion = 0.81,
  stem_distance = 100)

overlayGOValues(plot, goframe, label.terms = F, ...)

GO_parents(node = "GO:0008150", ontology = "BP")

GO_children(node = "GO:0008150", ontology = "BP")

tableGOTerms(organism, uniprot, wanted = c("GO:0030054", "GO:0005886",
  "GO:0012505", "GO:0005783", "GO:0005794", "GO:0005764", "GO:0016020",
  "GO:0005739", "GO:0005634", "GO:0005576", "GO:0005773", "GO:0005829"),
  ontology = "CC")
}
\description{
Deprecated functions

Get domain sets - Deprecated
}

