% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rnaseq.edgeR.getBCV}
\alias{rnaseq.edgeR.getBCV}
\title{Get the biological coefficient of variation for a set of reads}
\usage{
rnaseq.edgeR.getBCV(all.reads, filtered = T, ...)
}
\arguments{
\item{all.reads}{Reads as read in by \code{\link{rnaseq.readExpression}}}

\item{filtered}{Toggle if there should be a filtering for minimum number of reads (default TRUE)}
}
\value{
The biological coefficient of variation
}
\description{
Get the biological coefficient of variation for a set of reads
}

