% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rnaseq.getDifferential}
\alias{rnaseq.getDifferential}
\title{Perform a differential analysis on a set of expression values from RNAseq experiments}
\usage{
rnaseq.getDifferential(all.reads, filtered = T,
  clonal.variation.threshold = 1, pval.edger = 0.05, pval.deseq = 0.1,
  ...)
}
\arguments{
\item{all.reads}{Reads as read in by \code{\link{rnaseq.readExpression}}}

\item{filtered}{Toggle if there should be a filtering for minimum number of reads (default TRUE)}

\item{clonal.variation.threshold}{Threshold to drop genes if there is a log2 variation between clones. Set to NA to remove this filtering}

\item{pval.edger}{EdgeR adjusted p-value cutoff for differential expression}

\item{pval.deseq}{DESeq adjusted p-value cutoff for differential expression}
}
\value{
Data.frame with geneid logFC PValue uniprot and genename
}
\description{
Perform a differential analysis on a set of expression values from RNAseq experiments
}

