% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnaseq.readExpression}
\alias{rnaseq.readExpression}
\title{Read an expression file obtained from BGI}
\usage{
rnaseq.readExpression(filename)
}
\value{
Data frame with expression data
}
\description{
Read an expression file obtained from BGI
}

